+++ src/Entity/Item.php
@@ -182,5 +182,4 @@ class Item extends Base
     {
         return $this->getEntityInformations()->bonusSummary;
     }
-
 }
